requirements.txt
aiogram==2.25
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.3.0
Django==5.0.1
djangorestframework==3.14.0
frozenlist==1.4.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.2.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
magic-filter==1.0.12
multidict==6.0.4
pillow==10.2.0
pytz==2024.1
rfc3986==1.5.0
sniffio==1.3.0
sqlparse==0.4.4
tzdata==2023.4
yarl==1.9.2



















































































































































































